version set from package.json and release bin names
