+++ platformio.ini
@@ -151,10 +151,10 @@ build_flags = -g
   -DCONFIG_LITTLEFS_FOR_IDF_3_2
 
 [scripts_defaults]
-extra_scripts             = pio-scripts/name-firmware.py
-                            pio-scripts/gzip-firmware.py
-                            pio-scripts/strip-floats.py
-                            pio-scripts/user_config_copy.py
+extra_scripts = pre:pio-scripts/set_version.py
+  post:pio-scripts/output_bins.py
+  post:pio-scripts/strip-floats.py
+  pre:pio-scripts/user_config_copy.py
 
 # ------------------------------------------------------------------------------
 # COMMON SETTINGS:
@@ -213,7 +213,7 @@ platform = ${common.platform_wled_default}
 platform_packages = ${common.platform_packages}
 board_build.ldscript = ${common.ldscript_4m1m}
 build_unflags = ${common.build_unflags}
-build_flags = ${common.build_flags_esp8266}
+build_flags = ${common.build_flags_esp8266} -D WLED_RELEASE_NAME=ESP8266
 
 [env:esp01_1m_full]
 board = esp01_1m
@@ -221,7 +221,7 @@ platform = ${common.platform_wled_default}
 platform_packages = ${common.platform_packages}
 board_build.ldscript = ${common.ldscript_1m128k}
 build_unflags = ${common.build_unflags}
-build_flags = ${common.build_flags_esp8266} -D WLED_DISABLE_OTA
+build_flags = ${common.build_flags_esp8266} -D WLED_RELEASE_NAME=ESP01 -D WLED_DISABLE_OTA
 
 [env:esp07]
 board = esp07
@@ -261,7 +261,7 @@ build_flags = ${common.build_flags_esp8266} -D LEDPIN=1 -D WLED_DISABLE_INFRARED
 board = esp32dev
 platform = espressif32@3.2
 build_unflags = ${common.build_unflags}
-build_flags = ${common.build_flags_esp32}
+build_flags = ${common.build_flags_esp32} -D WLED_RELEASE_NAME=ESP32
 lib_ignore =
   ESPAsyncTCP
   ESPAsyncUDP
@@ -271,7 +271,7 @@ board = esp32-poe
 platform = espressif32@3.2
 upload_speed = 921600
 build_unflags = ${common.build_unflags}
-build_flags = ${common.build_flags_esp32} -D RLYPIN=-1 -D WLED_USE_ETHERNET -D BTNPIN=-1
+build_flags = ${common.build_flags_esp32} -D WLED_RELEASE_NAME=ESP32_Ethernet -D RLYPIN=-1 -D WLED_USE_ETHERNET -D BTNPIN=-1
 lib_ignore =
   ESPAsyncTCP
   ESPAsyncUDP
